refactor: reduce duplicate run template action code
