+++ tests/unit/action/test_run_job_template.py
@@ -106,11 +106,11 @@ async def test_run_job_template_exception(err_msg, err):
 
 DROOLS_CALLS = [
     (
-        "ansible_rulebook.action.run_job_template.lang.assert_fact",
+        "drools.ruleset.assert_fact",
         dict(set_facts=True),
     ),
     (
-        "ansible_rulebook.action.run_job_template.lang.post",
+        "drools.ruleset.post",
         dict(post_events=True),
     ),
 ]
@@ -209,7 +209,7 @@ async def test_run_job_template_retries():
         side_effect=controller_job,
     ):
         with patch(
-            "ansible_rulebook.action.run_job_template.lang.assert_fact"
+            "drools.ruleset.assert_fact",
         ) as drools_mock:
             await RunJobTemplate(metadata, control, **action_args)()
             drools_mock.assert_called_once()
